Moved RequirementProblem to int pkg
