+++ andhow/src/main/java/yarnandtail/andhow/internal/RequirementProblem.java
@@ -1,5 +1,8 @@
-package yarnandtail.andhow;
+package yarnandtail.andhow.internal;
 
+import yarnandtail.andhow.Problem;
+import yarnandtail.andhow.Property;
+import yarnandtail.andhow.PropertyGroup;
 import yarnandtail.andhow.util.TextUtil;
 
 /**
