+++ andhow/src/main/java/yarnandtail/andhow/util/ReportGenerator.java
@@ -9,7 +9,7 @@ import yarnandtail.andhow.Loader;
 import yarnandtail.andhow.internal.LoaderProblem;
 import yarnandtail.andhow.Property;
 import yarnandtail.andhow.PropertyGroup;
-import yarnandtail.andhow.RequirementProblem;
+import yarnandtail.andhow.internal.RequirementProblem;
 import yarnandtail.andhow.ValueProblem;
 import yarnandtail.andhow.ConstructionDefinition;
 import yarnandtail.andhow.SamplePrinter;
